refactor: add CausesSelectSection to CausesPage
